Enhance singly linked list with tail and length support, plus tests
- Added `tail` pointer to SLL for O(1) tail insertions
- Added `length` field to SLL for O(1) size retrieval and bounds checking
- Updated implementation:
  - `sll_add_head`, `sll_add_tail`, `sll_add_at_index` now increment length
  - `sll_delete_at_index` decrements length and updates tail when needed
  - `sll_get` now uses length for bounds validation
- Updated test suite:
  - Added assertions for correct tail updates
  - Added assertions for correct length updates
  - Covered edge cases (insert/delete at head, tail, empty list)
- Improved API consistency and safety
- Update Makefile

+++ include/dynamic_array.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <stddef.h>
 typedef struct {
     int *data;
     size_t length;    
@@ -10,8 +10,8 @@ void init_array(DynamicArray *arr, int capacity);
 void resize_array(DynamicArray *arr);
 void push(DynamicArray *arr, int value);
 void pop(DynamicArray *arr);
-int get(DynamicArray *arr, int index);
-void set(DynamicArray *arr, int index, int value);
+int get(DynamicArray *arr, size_t index);
+void set(DynamicArray *arr, size_t index, int value);
 void free_array(DynamicArray *arr);
 int find_index(DynamicArray *arr, int target); // works on unsorted arrays
 int get_index(DynamicArray *arr, int target);  // binary search, requires sorted array
